+++ portal-web/test/com/liferay/portalweb/portlet/blogs/blogsentry/addblogsentry/AddBlogsEntryTests.html
@@ -12,6 +12,12 @@
 <tr><td><a href="../../../../../portalweb/portlet/blogs/portlet/addportletblogs/AddPortletBlogsTest.html">AddPortletBlogsTest</a></td></tr>
 <tr><td><a href="AddBlogsEntryTest.html">AddBlogsEntryTest</a></td></tr>
 <tr><td><a href="ViewBlogsEntryTest.html">ViewBlogsEntryTest</a></td></tr>
+<tr><td><a href="ViewBlogsEntryTest.html">ViewBlogsEntry1Test</a></td></tr>
+<tr><td><a href="ViewBlogsEntryTest.html">ViewBlogsEntry2Test</a></td></tr>
+<tr><td><a href="ViewBlogsEntryTest.html">ViewBlogsEntry3Test</a></td></tr>
+<tr><td><a href="ViewBlogsEntryTest.html">ViewBlogsEntryDetailsTest</a></td></tr>
+<tr><td><a href="ViewBlogsEntryTest.html">ViewBlogsEntryRSSTest</a></td></tr>
+<tr><td><a href="ViewBlogsEntryTest.html">ViewBlogsEntryViewCountDetailsTest</a></td></tr>
 <tr><td><a href="TearDownBlogsEntryTest.html">TearDownBlogsEntryTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portal/util/TearDownPageTest.html">TearDownPageTest</a></td></tr>
 </tbody></table>
