+++ portal-web/test/com/liferay/portalweb/portlet/blogs/blogsentry/addblogsentry/ViewBlogsEntryDetailsTest.html
