+++ portal-web/test/com/liferay/portalweb/portlet/blogs/blogsentry/addblogsentry/ViewBlogsEntryRSSTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portalweb.portlet.blogs.blogsentry.viewblogsentrydetails;
+package com.liferay.portalweb.portlet.blogs.blogsentry.addblogsentry;
 
 import com.liferay.portalweb.portal.BaseTestCase;
 import com.liferay.portalweb.portal.util.RuntimeVariables;
@@ -20,8 +20,8 @@ import com.liferay.portalweb.portal.util.RuntimeVariables;
 /**
  * @author Brian Wing Shun Chan
  */
-public class ViewBlogsEntryDetailsTest extends BaseTestCase {
-	public void testViewBlogsEntryDetails() throws Exception {
+public class ViewBlogsEntryRSSTest extends BaseTestCase {
+	public void testViewBlogsEntryRSS() throws Exception {
 		selenium.open("/web/guest/home/");
 		loadRequiredJavaScriptModules();
 
@@ -45,29 +45,33 @@ public class ViewBlogsEntryDetailsTest extends BaseTestCase {
 			RuntimeVariables.replace("Blogs Test Page"));
 		selenium.waitForPageToLoad("30000");
 		loadRequiredJavaScriptModules();
-		selenium.clickAt("//div[@class='entry-title']/h2/a",
+		assertEquals(RuntimeVariables.replace("Blogs Entry Title"),
+			selenium.getText("//div[@class='entry-title']/h2/a"));
+		assertEquals(RuntimeVariables.replace("Blogs Entry Content"),
+			selenium.getText("//div[@class='entry-body']/p"));
+		assertEquals(RuntimeVariables.replace("RSS (Opens New Window)"),
+			selenium.getText("//div[@class='subscribe']/span[1]/a"));
+
+		String rssURL = selenium.getAttribute(
+				"//div[@class='subscribe']/span[1]/a@href");
+		RuntimeVariables.setValue("rssURL", rssURL);
+		selenium.open(RuntimeVariables.getValue("rssURL"));
+		loadRequiredJavaScriptModules();
+		assertEquals(RuntimeVariables.replace("Liferay"),
+			selenium.getText("//x:h1[@id='feedTitleText']"));
+		assertEquals(RuntimeVariables.replace("Liferay"),
+			selenium.getText("//x:h2[@id='feedSubtitleText']"));
+		assertEquals(RuntimeVariables.replace("Blogs Entry Title"),
+			selenium.getText("//x:div[@id='feedContent']/x:div[1]/x:h3/x:a"));
+		assertEquals(RuntimeVariables.replace("Blogs Entry Content"),
+			selenium.getText("//x:div[@id='feedContent']/x:div[1]/x:div/x:p"));
+		selenium.clickAt("//x:div[@id='feedContent']/x:div[1]/x:h3/x:a",
 			RuntimeVariables.replace("Blogs Entry Title"));
 		selenium.waitForPageToLoad("30000");
 		loadRequiredJavaScriptModules();
 		assertEquals(RuntimeVariables.replace("Blogs Entry Title"),
 			selenium.getText("//h1[@class='header-title']/span"));
 		assertEquals(RuntimeVariables.replace("Blogs Entry Content"),
-			selenium.getText("//div[@class='entry-body']"));
-		assertEquals(RuntimeVariables.replace("By Joe Bloggs"),
-			selenium.getText("//div[@class='entry-author']"));
-		assertEquals(RuntimeVariables.replace("0 Comments"),
-			selenium.getText("//span[@class='comments']"));
-		assertEquals(RuntimeVariables.replace("Your Rating"),
-			selenium.getText(
-				"xPath=(//div[@class='aui-rating-label-element'])[1]"));
-		assertEquals(RuntimeVariables.replace("Average (0 Votes)"),
-			selenium.getText(
-				"xPath=(//div[@class='aui-rating-label-element'])[2]"));
-		assertEquals(RuntimeVariables.replace("Previous"),
-			selenium.getText("//span[@class='previous']"));
-		assertEquals(RuntimeVariables.replace("Next"),
-			selenium.getText("//span[@class='next']"));
-		assertEquals(RuntimeVariables.replace("Comments"),
-			selenium.getText("//div[@class='lfr-panel-title']/span"));
+			selenium.getText("//div[@class='entry-body']/p"));
 	}
 }
