+++ portal-web/test/com/liferay/portalweb/portlet/blogs/blogsentry/deleteblogsentrytitleescapecharacters/DeleteBlogsEntryTitleEscapeCharactersTests.html
@@ -10,7 +10,7 @@
 <tr><td><b>DeleteBlogsEntryTitleEscapeCharactersTests</b></td></tr>
 <tr><td><a href="../../../../../portalweb/portlet/blogs/portlet/addportletblogs/AddPageBlogsTest.html">AddPageBlogsTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portlet/blogs/portlet/addportletblogs/AddPortletBlogsTest.html">AddPortletBlogsTest</a></td></tr>
-<tr><td><a href="AddBlogsEntryTitleEscapeCharactersTest.html">AddBlogsEntryTitleEscapeCharactersTest</a></td></tr>
+<tr><td><a href="../../../../../portalweb/portlet/blogs/blogsentry/addblogsentrytitleescapecharacters/AddBlogsEntryTitleEscapeCharactersTest.html">AddBlogsEntryTitleEscapeCharactersTest</a></td></tr>
 <tr><td><a href="DeleteBlogsEntryTitleEscapeCharactersTest.html">DeleteBlogsEntryTitleEscapeCharactersTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portlet/blogs/blogsentry/addblogsentry/TearDownBlogsEntryTest.html">TearDownBlogsEntryTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portal/util/TearDownPageTest.html">TearDownPageTest</a></td></tr>
