+++ portal-web/test/com/liferay/portalweb/portlet/blogs/blogsentry/deleteblogsentrytitleescapecharacters/DeleteBlogsEntryTitleEscapeCharactersTests.java
@@ -17,6 +17,7 @@ package com.liferay.portalweb.portlet.blogs.blogsentry.deleteblogsentrytitleesca
 import com.liferay.portalweb.portal.BaseTestSuite;
 import com.liferay.portalweb.portal.util.TearDownPageTest;
 import com.liferay.portalweb.portlet.blogs.blogsentry.addblogsentry.TearDownBlogsEntryTest;
+import com.liferay.portalweb.portlet.blogs.blogsentry.addblogsentrytitleescapecharacters.AddBlogsEntryTitleEscapeCharactersTest;
 import com.liferay.portalweb.portlet.blogs.portlet.addportletblogs.AddPageBlogsTest;
 import com.liferay.portalweb.portlet.blogs.portlet.addportletblogs.AddPortletBlogsTest;
