--- portal-web/test/com/liferay/portalweb/portlet/blogs/blogsentry/viewblogsentryviewcountdetails/ViewBlogsEntryViewCountDetailsTests.html
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en" lang="en">
-<head>
-  <meta content="text/html; charset=UTF-8" http-equiv="content-type" />
-  <title>ViewBlogsEntryViewCountDetailsTests</title>
-</head>
-<body>
-<table id="suiteTable" cellpadding="1" cellspacing="1" border="1" class="selenium"><tbody>
-<tr><td><b>ViewBlogsEntryViewCountDetailsTests</b></td></tr>
-<tr><td><a href="../../../../../portalweb/portlet/blogs/portlet/addportletblogs/AddPageBlogsTest.html">AddPageBlogsTest</a></td></tr>
-<tr><td><a href="../../../../../portalweb/portlet/blogs/portlet/addportletblogs/AddPortletBlogsTest.html">AddPortletBlogsTest</a></td></tr>
-<tr><td><a href="../../../../../portalweb/portlet/blogs/blogsentry/addblogsentry/AddBlogsEntryTest.html">AddBlogsEntryTest</a></td></tr>
-<tr><td><a href="ViewBlogsEntryViewCountDetailsTest.html">ViewBlogsEntryViewCountDetailsTest</a></td></tr>
-<tr><td><a href="../../../../../portalweb/portlet/blogs/blogsentry/addblogsentry/TearDownBlogsEntryTest.html">TearDownBlogsEntryTest</a></td></tr>
-<tr><td><a href="../../../../../portalweb/portal/util/TearDownPageTest.html">TearDownPageTest</a></td></tr>
-</tbody></table>
-</body>
-</html>
